refactor core package and add 'net generator' interface
