+++ core/core.go
@@ -7,7 +7,6 @@ import (
 
 	context "github.com/jbenet/go-ipfs/Godeps/_workspace/src/code.google.com/p/go.net/context"
 	b58 "github.com/jbenet/go-ipfs/Godeps/_workspace/src/github.com/jbenet/go-base58"
-	ctxgroup "github.com/jbenet/go-ipfs/Godeps/_workspace/src/github.com/jbenet/go-ctxgroup"
 	datastore "github.com/jbenet/go-ipfs/Godeps/_workspace/src/github.com/jbenet/go-datastore"
 	ma "github.com/jbenet/go-ipfs/Godeps/_workspace/src/github.com/jbenet/go-multiaddr"
 
@@ -26,20 +25,14 @@ import (
 	dht "github.com/jbenet/go-ipfs/routing/dht"
 	offroute "github.com/jbenet/go-ipfs/routing/offline"
 
-	bstore "github.com/jbenet/go-ipfs/blocks/blockstore"
-	bserv "github.com/jbenet/go-ipfs/blockservice"
-	exchange "github.com/jbenet/go-ipfs/exchange"
 	bitswap "github.com/jbenet/go-ipfs/exchange/bitswap"
 	bsnet "github.com/jbenet/go-ipfs/exchange/bitswap/network"
-	offline "github.com/jbenet/go-ipfs/exchange/offline"
 	rp "github.com/jbenet/go-ipfs/exchange/reprovide"
 
 	mount "github.com/jbenet/go-ipfs/fuse/mount"
 	merkledag "github.com/jbenet/go-ipfs/merkledag"
 	namesys "github.com/jbenet/go-ipfs/namesys"
 	path "github.com/jbenet/go-ipfs/path"
-	pin "github.com/jbenet/go-ipfs/pin"
-	repo "github.com/jbenet/go-ipfs/repo"
 	config "github.com/jbenet/go-ipfs/repo/config"
 )
 
@@ -58,40 +51,6 @@ const (
 	onlineMode
 )
 
-// IpfsNode is IPFS Core module. It represents an IPFS instance.
-type IpfsNode struct {
-
-	// Self
-	Identity peer.ID // the local node's identity
-
-	Repo repo.Repo
-
-	// Local node
-	Pinning    pin.Pinner // the pinning manager
-	Mounts     Mounts     // current mount state, if any.
-	PrivateKey ic.PrivKey // the local node's private Key
-
-	// Services
-	Peerstore  peer.Peerstore       // storage for other Peer instances
-	Blockstore bstore.Blockstore    // the block store (lower level)
-	Blocks     *bserv.BlockService  // the block service, get/add blocks.
-	DAG        merkledag.DAGService // the merkle dag service, get/add objects.
-	Resolver   *path.Resolver       // the path resolution system
-
-	// Online
-	PeerHost     p2phost.Host        // the network host (server+client)
-	Bootstrapper io.Closer           // the periodic bootstrapper
-	Routing      routing.IpfsRouting // the routing system. recommend ipfs-dht
-	Exchange     exchange.Interface  // the block exchange + strategy (bitswap)
-	Namesys      namesys.NameSystem  // the name system, resolves paths to hashes
-	Diagnostics  *diag.Diagnostics   // the diagnostics service
-	Reprovider   *rp.Reprovider      // the value reprovider system
-
-	ctxgroup.ContextGroup
-
-	mode mode
-}
-
 // Mounts defines what the node's mount state is. This should
 // perhaps be moved to the daemon or mount. It's here because
 // it needs to be accessible across daemon requests.
@@ -100,122 +59,7 @@ type Mounts struct {
 	Ipns mount.Mount
 }
 
-type ConfigOption func(ctx context.Context) (*IpfsNode, error)
-
-func NewIPFSNode(parent context.Context, option ConfigOption) (*IpfsNode, error) {
-	ctxg := ctxgroup.WithContext(parent)
-	ctx := ctxg.Context()
-	success := false // flip to true after all sub-system inits succeed
-	defer func() {
-		if !success {
-			ctxg.Close()
-		}
-	}()
-
-	node, err := option(ctx)
-	if err != nil {
-		return nil, err
-	}
-	node.ContextGroup = ctxg
-	ctxg.SetTeardown(node.teardown)
-
-	// Need to make sure it's perfectly clear 1) which variables are expected
-	// to be initialized at this point, and 2) which variables will be
-	// initialized after this point.
-
-	node.Blocks, err = bserv.New(node.Blockstore, node.Exchange)
-	if err != nil {
-		return nil, debugerror.Wrap(err)
-	}
-	if node.Peerstore == nil {
-		node.Peerstore = peer.NewPeerstore()
-	}
-	node.DAG = merkledag.NewDAGService(node.Blocks)
-	node.Pinning, err = pin.LoadPinner(node.Repo.Datastore(), node.DAG)
-	if err != nil {
-		node.Pinning = pin.NewPinner(node.Repo.Datastore(), node.DAG)
-	}
-	node.Resolver = &path.Resolver{DAG: node.DAG}
-	success = true
-	return node, nil
-}
-
-func Offline(r repo.Repo) ConfigOption {
-	return Standard(r, false)
-}
-
-func OnlineWithRouting(r repo.Repo, router routing.IpfsRouting) ConfigOption {
-	if router == nil {
-		panic("router required")
-	}
-	return standardWithRouting(r, true, router)
-}
-
-func Online(r repo.Repo) ConfigOption {
-	return Standard(r, true)
-}
-
-// DEPRECATED: use Online, Offline functions
-func Standard(r repo.Repo, online bool) ConfigOption {
-	return standardWithRouting(r, online, nil)
-}
-
-// TODO refactor so maybeRouter isn't special-cased in this way
-func standardWithRouting(r repo.Repo, online bool, maybeRouter routing.IpfsRouting) ConfigOption {
-	return func(ctx context.Context) (n *IpfsNode, err error) {
-		// FIXME perform node construction in the main constructor so it isn't
-		// necessary to perform this teardown in this scope.
-		success := false
-		defer func() {
-			if !success && n != nil {
-				n.teardown()
-			}
-		}()
-
-		// TODO move as much of node initialization as possible into
-		// NewIPFSNode. The larger these config options are, the harder it is
-		// to test all node construction code paths.
-
-		if r == nil {
-			return nil, debugerror.Errorf("repo required")
-		}
-		n = &IpfsNode{
-			mode: func() mode {
-				if online {
-					return onlineMode
-				}
-				return offlineMode
-			}(),
-			Repo: r,
-		}
-
-		// setup Peerstore
-		n.Peerstore = peer.NewPeerstore()
-
-		// setup local peer ID (private key is loaded in online setup)
-		if err := n.loadID(); err != nil {
-			return nil, err
-		}
-
-		n.Blockstore, err = bstore.WriteCached(bstore.NewBlockstore(n.Repo.Datastore()), kSizeBlockstoreWriteCache)
-		if err != nil {
-			return nil, debugerror.Wrap(err)
-		}
-
-		if online {
-			if err := n.startOnlineServices(ctx, maybeRouter); err != nil {
-				return nil, err
-			}
-		} else {
-			n.Exchange = offline.Exchange(n.Blockstore)
-		}
-
-		success = true
-		return n, nil
-	}
-}
-
-func (n *IpfsNode) startOnlineServices(ctx context.Context, maybeRouter routing.IpfsRouting) error {
+func (n *IpfsNode) startOnlineServices(ctx context.Context, maybeRouter routing.IpfsRouting, nso NetworkSetupOption) error {
 
 	if n.PeerHost != nil { // already online.
 		return debugerror.New("node already online")
@@ -226,7 +70,8 @@ func (n *IpfsNode) startOnlineServices(ctx context.Context, maybeRouter routing.
 		return err
 	}
 
-	peerhost, err := constructPeerHost(ctx, n.Identity, n.Peerstore)
+	// Construct peer host from option
+	peerhost, err := nso(ctx, n.Identity, n.Peerstore)
 	if err != nil {
 		return debugerror.Wrap(err)
 	}
