+++ core/mock.go
@@ -6,13 +6,17 @@ import (
 	syncds "github.com/jbenet/go-ipfs/Godeps/_workspace/src/github.com/jbenet/go-datastore/sync"
 	"github.com/jbenet/go-ipfs/blocks/blockstore"
 	blockservice "github.com/jbenet/go-ipfs/blockservice"
+	bitswap "github.com/jbenet/go-ipfs/exchange/bitswap"
+	bsnet "github.com/jbenet/go-ipfs/exchange/bitswap/network"
 	"github.com/jbenet/go-ipfs/exchange/offline"
 	mdag "github.com/jbenet/go-ipfs/merkledag"
 	nsys "github.com/jbenet/go-ipfs/namesys"
+	gen "github.com/jbenet/go-ipfs/p2p/net/gen"
 	mocknet "github.com/jbenet/go-ipfs/p2p/net/mock"
 	peer "github.com/jbenet/go-ipfs/p2p/peer"
 	path "github.com/jbenet/go-ipfs/path"
 	"github.com/jbenet/go-ipfs/repo"
+	dht "github.com/jbenet/go-ipfs/routing/dht"
 	mockrouting "github.com/jbenet/go-ipfs/routing/mock"
 	ds2 "github.com/jbenet/go-ipfs/util/datastore2"
 	testutil "github.com/jbenet/go-ipfs/util/testutil"
@@ -57,6 +61,64 @@ func NewMockNode() (*IpfsNode, error) {
 	// Bitswap
 	bstore := blockstore.NewBlockstore(nd.Repo.Datastore())
 	bserv, err := blockservice.New(bstore, offline.Exchange(bstore))
+	if err != nil {
+		return nil, err
+	}
+
+	nd.DAG = mdag.NewDAGService(bserv)
+
+	// Namespace resolver
+	nd.Namesys = nsys.NewNameSystem(nd.Routing)
+
+	// Path resolver
+	nd.Resolver = &path.Resolver{DAG: nd.DAG}
+
+	return nd, nil
+}
+
+func NewNodeOnNetwork(ngen gen.NetGenerator) (*IpfsNode, error) {
+	ctx := context.TODO()
+	nd := new(IpfsNode)
+
+	// Generate Identity
+	ident, err := testutil.RandIdentity()
+	if err != nil {
+		return nil, err
+	}
+
+	p := ident.ID()
+	nd.Identity = p
+	nd.PrivateKey = ident.PrivateKey()
+	nd.Peerstore = peer.NewPeerstore()
+	nd.Peerstore.AddPrivKey(p, ident.PrivateKey())
+	nd.Peerstore.AddPubKey(p, ident.PublicKey())
+
+	ph, err := ngen.AddPeer(ident.PrivateKey(), ident.ID())
+	if err != nil {
+		return nil, err
+	}
+
+	nd.PeerHost = ph
+	if err != nil {
+		return nil, err
+	}
+
+	// Temp Datastore
+	nd.Repo = &repo.Mock{
+		// TODO C: conf,
+		D: ds2.CloserWrap(syncds.MutexWrap(datastore.NewMapDatastore())),
+	}
+
+	// Routing
+	nd.Routing = dht.NewDHT(ctx, ph, nd.Repo.Datastore())
+
+	// Bitswap
+	bstore := blockstore.NewBlockstore(nd.Repo.Datastore())
+
+	bitnet := bsnet.NewFromIpfsHost(nd.PeerHost, nd.Routing)
+	bswap := bitswap.New(ctx, nd.Identity, bitnet, bstore, true)
+
+	bserv, err := blockservice.New(bstore, bswap)
 	if err != nil {
 		return nil, err
 	}
